+++ src/idunno.Bluesky/Record/BlueskyRecord.cs
@@ -4,6 +4,7 @@
 using System.Text.Json.Serialization;
 
 using idunno.AtProto.Repo;
+using idunno.Bluesky.Actor;
 using idunno.Bluesky.Feed;
 using idunno.Bluesky.Graph;